requirements.txt
colorama==0.3.9
oauthlib==2.0.2
py-trello==0.9.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
